+++ apps/server/src/shared/domain/interface/index.ts
@@ -6,4 +6,3 @@ export * from './permission';
 export * from './permission.enum';
 export * from './rolename.enum';
 export * from './user';
-export * from './hasPermission';
